% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Generates a map displaying accidents for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer}

\item{year}{integer}
}
\value{
This function returns a graphical object.
}
\description{
This is a function that generates a map (using the maps package) that displays the location of
accidents (using the graphics package) for a given state and year.
}
\details{
Will generate error if an invalid state number is input.
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
library(maps)
library(graphics)
f14path <- system.file("extdata", "accident_2014.csv.bz2", package = "farsfuncs")
file.copy(from=c(f14path),to=getwd())
fars_map_state(1,2014)

}
