% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{fars_read}
\alias{fars_read}
\title{Read fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{string}
}
\value{
This function returns a table data frame.
}
\description{
This is a simple function that reads a csv data file into R.
}
\details{
If the file does not exist it will generate an error message.
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
f13path<-system.file("extdata", "accident_2013.csv.bz2", package = "farsfuncs")
file.copy(from=c(f13path),to=getwd())
fars_read("accident_2013.csv.bz2")

}
