% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generates a table showing number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{integer}
}
\value{
A table showing total number of accident by month and year.
}
\description{
Function to generate a table showing numbers of accidents for a list of years from fars data.
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
f13path <- system.file("extdata", "accident_2013.csv.bz2", package = "farsfuncs")
f14path <- system.file("extdata", "accident_2014.csv.bz2", package = "farsfuncs")
f15path <- system.file("extdata", "accident_2015.csv.bz2", package = "farsfuncs")
file.copy(from=c(f13path,f14path,f15path),to=getwd())
fars_summarize_years(c(2013,2014,2015))

}
