% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{make_filename}
\alias{make_filename}
\title{Generates a fars file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer}
}
\value{
String that represents a fars file name for a given year.
}
\description{
Simple function to generate a character string that is in the standard form for a fars data file.
}
\examples{
make_filename(2017)

}
